+++ CHANGELOG.md
@@ -1,8 +1,15 @@
-# Change Log
+# Changelog
+
+## v1.4.0
+
+- Update dependencies
+- Remove flow
+- Remove Travis CI config
+- Ensure webpack 5 support
 
 ## v1.1.0
 
- - Added `min` and `max` options to automatically generate a number of images, and `steps` option to say how many images ([#31](https://github.com/herrstucki/responsive-loader/pull/31)).
+- Added `min` and `max` options to automatically generate a number of images, and `steps` option to say how many images ([#31](https://github.com/herrstucki/responsive-loader/pull/31)).
 
 ## v1.0.0